Extract app view models into dedicated package
Move UI state models to app.view.model subpackage for cleaner
separation between composables and data classes.
