+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/view/model/MemosScreen.kt
@@ -1,0 +1,7 @@
+package space.be1ski.vibits.shared.app.view.model
+
+internal enum class MemosScreen {
+  HABITS,
+  STATS,
+  FEED,
+}
